Create tool to dump information about all settings
Summary:
I'm creating a tool that will dump information about all settings, both client and server into JSON format.

The information contains information such as the name of the setting, the name of the bundle, the default value, whether it's a server or client setting, whether it's experimental, deprecated, internal, etc.

Reviewed By: MohamedBassem

Differential Revision: D22137530

fbshipit-source-id: b61b363cebec210231081fade9497ad9f63b5ea5