Now used SHARD_UNDRAIN when we transition from NONE instead of aborting rebuilding
Summary:
This addresses the case at which we have drain=1 and rebuilding was in RESTORE mode, the internal maintenance will be removed by the rebuilding complete metadata marker will not be written. In this workflow change, we will use the SHARD_UNDRAIN event instead of just aborting rebuilding. This event will (also) abort rebuilding but will trigger the marker write on the storage node and the shard rebuilding state will be in-sync in this scenario.

This is a short-term fix for a problem that shouldn't exist. Ideally what we need is:
1- Get rid of drain flag in authoritative status. The only mechanism to switch from EMPTY -> FULLY_AUTH should be via MM.
2- No need for authoritative status altogether. When rebuilding is done, the node will transition into NONE, the node will write its own marker once its rebuilding is acked/finished. Then the rest of the workflow will be handled by MM. The local update of the shard state should not directly mean that the shard is FULLY_AUTHORTATIVE.

These changes to RebuildingCoordinator (and recoveries and client read stream) will be done as a separate project.

Reviewed By: MohamedBassem

Differential Revision: D22356467

fbshipit-source-id: 4d439d78d8f79bf97e48c7935bf818caf7967a37