Enable Hazptr thread pool executor
Summary:
LogDevice doesn't call `folly::init` which means it misses the initialization of the hazptr reclamation thread pool executor. The executor defaults to an InlineExecutor. This can lead to an insertion into a ConcurrentHashMap resulting into a reclamation of another object in another CHM inline. This can deadlock if both the insertions and destructions are synchronized on the same lock.

The correct fix is to start adopting `folly::init` but that's a more disruptive change for a later diff.

Reviewed By: tau0

Differential Revision: D22501213

fbshipit-source-id: b08c8ac764a86c4a1360918175013b5db9c4c246