Deprecating TimeoutMap
Summary:
It is a legacy from pre folly world, which was required to speed up timer scheduling. In the new world the better approach to use HHWheelTimer for timers which are scheduled often.

Since the migration to folly it hasn't been used. So it is time to remove it from codebase.

Reviewed By: 747mmHg

Differential Revision: D22552964

fbshipit-source-id: 8c414b8f22e4c5b1624596ff07de67ef8f9b958c