Adding support for the new nodes configuration in Configuration.h
Summary:
In preparation for the old nodes configuration (based on server config) deprecation, I'm adding support for the new nodes configuration in the `Configuration` struct and doing the needed updates in the `UpdatableConfig` class as well.

Optimally, I'd want to get rid of the `Configuration` class but that's a different story.

Reviewed By: tau0

Differential Revision: D22505629

fbshipit-source-id: c27c5464a8bc5bde69564b57d53334e7191f47ee