Migratiing users of NodeSetTestUtil to the new nodes configuration structure
Summary:
The first major migration for users of `getNodesConfigurationFromServerConfigSource`. This diff migrates `NodeSetTestUtil` that's used in a lot of tests for creating nodes config to using the nodes configuration structure. This required modification to every single callsite to change the passed argument into being a NodesConfiguration instead of NodesConfig.

The changes are mostly:
- Replace the passed Nodes structure to a newly created NodesConfig.
- Modify all the `addNodes` calls to pass the shared_ptr of the config by reference.
- If metadata changes are required, I apply updates on the nodes configuration structure.
- I modify the code that creates the ServerConfig out of this nodes config to not pass the nodes config and instead pass it to the `Configuration` structure directly.
- If NodesConfiguration is the only needed part of the config, I changed the test to use `NodesConfiguration` instead of the `Configuration` structure.

The diff is huge but it's mostly mechanical.

Reviewed By: tau0

Differential Revision: D22524898

fbshipit-source-id: 728d0e43973331785860f80097703e70b2b0b2dd