Support payload groups in buffered writer
Summary:
`BufferedWriter` API is changed to support writing `PayloadGroup`s. This is a breaking change due to `payload` field type change in `BufferedWriter::Append`, as payload can be either `std::string` or `PayloadGroup`. Another breaking change is in `BufferedWriter:: AppendCallback::ContextSet`, since payload ownership needs to be returned back to the caller. Alternative to this is adding alternatives to `Append` and `AppendCallback`, but this requires way more changes.

Migration to the new API (assuming caller keeps using string payloads) is straightforward and only involves adding `std::get<std::string>(...)` to access payload.

Reviewed By: tau0

Differential Revision: D22306499

fbshipit-source-id: f664000dc30a71cde198e7f5831513ddb2bdb297