Deprecate NodeTemplate and use configuration::Node with a builder instead
Summary:
This diff deprecates the `NodeTemplate` structure and uses the old `Node.h` structure instead. The main motivation behind this change is to make migrating the integration tests much easier as they rely heavily on the `Node.h` structure.

This diff also fixes oss build as it moves away for the designated Initializers and instead uses a builder.

Eventually, once all the legacy NodesConfig are deprecated, we can move the `Node.h` structure to a test directory.

Note: The diff is huge, but it's mostly mechanical changes.

Reviewed By: 747mmHg

Differential Revision: D22662284

fbshipit-source-id: 896aad0b3ad76b7fb1d9efeade39f4de21ea887c