Bugfix for maintenance log trimming.
Summary:
Fix a bug in `ReplicatedStateMachine` where `snapshot_in_flight_` was not set to false immediately after the snapshot callback. Implementations of virtual methods like `onSnapshotCreated` relied on this but were called before the flag was set to false, resulting in inconsistencies.

In particular, this caused `ClusterMaintenanceStateMachine` to never trim the snapshot log.

Reviewed By: AhmedSoliman

Differential Revision: D22921756

fbshipit-source-id: 3d02278a753fa7cf4584c9becf3ce3173575ab25