NodesConfiguration: Fix client tests that doesn't have a running cluster
Summary: In a world where NCM is the default, clients can't start without having a provisioned NCM. This diff introduces a new test utility that provisions an NodesConfigurationStore in a temp directory and return it to the caller. Tests that starts clients without a running cluster can utilize to build a client.

Reviewed By: 747mmHg

Differential Revision: D22950564

fbshipit-source-id: eb4019b68cb69e914d4a60cf9e44d9bb81fbd993