Migrate users of IntegrationTestUtils to use the NodesConfiguration structure
Summary:
Now that the integration testing framework is ready for using NCM everywhere, let's modify all the callistes to pass the appropriate nodes configuration. The diff is huge, but it's mostly mechanical changes with those rules:
1. If the caller uses a trivial nodes config, convert it into `createSimpleNodesConfig` helper function.
2. If the caller uses a `configuration::Nodes` structure, convert it into NCM using `NodesConfigurationTestUtil::provisionNodes` and take into account if there's any specific metadata replication needed.
3. If the caller uses a config from a file, use a `withNodesConfiguration` function on the `Configuration::fromJson` call.

After converting the initial NCM, apply the necessary changes (e.g. storage state changes) on the resulting NC before passing it to the test cluster.

Reviewed By: 747mmHg

Differential Revision: D22791456

fbshipit-source-id: cc8ebabfcd89f7e39532d3bc221a6fa5ee5eafb5