NodesConfiguration: Migrate users for Node.h functions to NodesConfiguration
Summary:
Convert function calls that uses the legacy NodesConfig to the corresponding NodesConfiguration calls. This is needed because the NodesConfig is going to be empty in most of the tests after the last diff in the stack.

Note: It might seem that I'm removing some constraints from `NodesConfigLegacyConverter.cpp` but this class is actually going to be completed removed in the next diff.

Reviewed By: 747mmHg

Differential Revision: D22846452

fbshipit-source-id: 8fb773274dbfc0484cd91b8cf1ab8d5889f8044e