Adding support for FORCE_RESTORE flag in eventlog and use it in maintenance manager
Summary:
## Background
Rebuilding has a separate control plane state machine (EventLog) that reacts to events like (`SHARD_NEEDS_REBUILD`, `SHARD_UNDRAIN`, etc.). There are multiple ways at which we trigger and abort rebuilding. a) through Maintenance requests (MM) b) through RebuildingSupervisor/RebuildingCoordinator.

## The `DRAIN` flag
Whenever we trigger rebuilding, we get to choose two things, whether this is a `DRAIN`, and whether we want to rebuild in `RELOCATE` or `RESTORE` mode. Without this diff, it's not possible to trigger a `DRAIN` rebuild without being in `RELOCATE`.

### What is the DRAIN flag anyway?
During rebuilding, we wait for all donors to finish rebuilding a shard before considering the authoritative status of a shard to be `AUTHORITATIVE_EMPTY`. The transition back to `FULLY_AUTHORITATIVE` means that the shard is back online and enabled. This used to happen automatically when the node starts up with its shards healthy. This is not true any more in maintenance manager workflow. What we want instead is to remove the internal maintenance and wait for maintenance manager to evaluate whether we want to enable the shard or not based on the state of the rest of the maintenances (other maintenances may still be applied on this node). We used the DRAIN flag in `RELOCATE` rebuilding to disable this automatic acknowledgement of rebuilding (transitioning to `FULLY_AUTHORITATIVE`). However, when the rebuilding runs in `RESTORE` mode, there is no way to also specify the flag `DRAIN`.

## Problem
The problem is that even that we use maintenance manager as the sole controller of rebuilding in production, we still keep the old code that may cancel rebuilding directly from RebuildingCoordinator in some corner cases. In principle, we want maintenance manager to be the **only** entity triggering or cancelling rebuilding (for full shard rebuilds). There are numerous corner cases in this code that can be cleaned up if we removed support for the old-style rebuilding trigger. I'm planning to do this incrementally.

In principle, with maintenance manager, ALL rebuilding requests should have the `DRAIN` flag since we don't want the machine to transition to `FULLY_AUTHORITATIVE` without being permitted to be enabled by maintenance manager.

## Plan
The end goal is to completely remove the need for the DRAIN flag since this will be the (always true) case. We also want to remove support for the direct removal/creation of rebuilding from rebuilding coordinator/supervisor. Instead, we want to have maintenance manager be the single controller. Rebuilding coordinator/supervisor will only create and remove internal maintenances and leave the rest to MM.

Steps to get there:
- I've added support in this diff for a new `FORCE_RESTORE` flag that forces to use RESTORE mode rebuilding and will work in conjunction with `DRAIN` flag. This makes it explicit if we want RESTORE or RELOCATE rebuilding instead of the implicit reliance on the DRAIN flag. (Currently, DRAIN means RELOCATE+no automatic rebuilding ack)
- In maintenance manager, we will set the FORCE_RESTORE (controlled behind a feature gating to allow us to safely roll this out) on restore rebuilding. MM will also set the DRAIN flag for these maintenances too.
- In the server side, we will respect the DRAIN flag even if in RESTORE mode, and FORCE_RESTORE will trigger a RESTORE rebuilding.

This is in preparation to remove the DRAIN flag altogether. Future diffs will incrementally transition into a state where this special casing is not needed and everything is a lot simpler to reason about.

Reviewed By: MohamedBassem

Differential Revision: D22844370

fbshipit-source-id: fd184a4d8d81b6d16aef72324e777ee4099f1ec6