EpochStoreV2: Drop support for reloading zookeeper config in epoch store
Summary: For the sake of simplification, let's drop support for live reloading of zookeeper config. It rarely changes and when it does, we can just restart the cluster so it's not worth the complexity. It's also one diff in the path of making `ZookeeperEpochStore` agnostic the underlying store and absorb the FileEpochStore implementation.

Reviewed By: 747mmHg

Differential Revision: D23186931

fbshipit-source-id: 94a80743b5dcd601ad1e560ee2b87aeb839f9b3e