Stats: fix PerMonitoringTagStats copy constructor
Summary:
There was an issue with some counters missing for PerMonitoringTagStats, and
the issue was the copy constructor not being correct.

Why would a copy constructor be an issue for us if we just call ::aggregate()
functions to propagate statistics data, you ask? The data is reshuffled in the
hash map we use, and, since PerMonitoringTagStats doesn't have a
move-constructor defined, the copy constructor is used as an alternative.

Reviewed By: MohamedBassem

Differential Revision: D23272904

fbshipit-source-id: cc6a641c50a91ba31d268a3679948f2822bbf8f8