Fix bug where client in low-level network priority could not startup
Summary:
There was an issue when starting up a client with "client-default-network-priority"="low" where NodesConfiguration did not contain low level priority during boostrapping.
This caused the client to be isolated as it could not use the default_client_data_address to connect and there was no low level priority address to use.

Reviewed By: MohamedBassem

Differential Revision: D23266837

fbshipit-source-id: 1072da99b4acdb99d454e7889a6ac653c96ae4ea