Introduce the RequestExecutor interface
Summary:
Abstracting the processor's `post*` APIs into an interface that's easily mockable in tests. Implementing the interface requires implementing a single function which is: Enqueue the given request, on the target worker. This diff also implements one possible mock that executes the requests inline.

Note: Some of the stuff had to be moved around to make a clean interface. For example, selection of execution thread and shutdown checks.

Reviewed By: tau0

Differential Revision: D23240681

fbshipit-source-id: 45caa9ac233df73cd914b9ab69fd7bd08951fc9a