ClientReadersFlowTracer: only publish traces after we update state (stuck/lagging)
Summary:
We should only publish traces immediately after assessing if the reader is stuck or lagging
because this allows for more accurate information (incl. information about the
current state, such as which nodes we are waiting for, which nodes are
unavailable, etc).

The only possible downside is that read speed might be calculated incorrectly.

Differential Revision: D21906139

fbshipit-source-id: 7edf1a3b4cd2be997f8b9e6901168caf1f987bab