Control where Thrift/Future client callbacks executed
Summary:
There are multiple ways of calling Thrift RPC on the client side with respect to how result of the call is delivered:
1. sync: blocking call, client thread is waiting.
2. async: asynchronous, result delivered through callback
3. Future-based: caller gets Future that is fulfilled when response is received
4. SemiFuture: the same as above but with SemiFuture instead of Future
5. coroutine (if enabled): caller gets coro task which is completed with RPC result

This abundance of APIs gives us a problem because on our side we need to integrate them somehow with our Worker-based threading model.
### How we will do it
#1 is trivial (if you call it from Worker you will get result in the same Worker thread but nonetheless you should never use it in production because it consumes the whole Worker).
#4 and #5 trivial as well because caller has to provide execution context for getting result anyway and can simply use Worker.
But #2 is problematic because by default it will deliver response in Thrift's own IO thread even if client method has been called from Worker (as well as #3 but this one should never be used in practice because Futures are deprecated).

### How to solve async methods?
Async methods require callback to passed. We will provide option to set callback executor when creating a new Thrift client and when async_ method is called we will wrap client-provided callback into one which simply delegates execution to this callback executor. As a nice side benefit It will work with Future methods as well.

Reviewed By: MohamedBassem

Differential Revision: D22866545

fbshipit-source-id: 781b36ba38506cd9a1c78162cd1127afab7ed39e