Restrict recovery writes to READ_WRITE nodes instead of nodes not in the rebuilding set
Summary:
Recoveries are not supposed to be writing to read only nodes. Before this diff, a shard is not considered writable only if it's in the rebuilding set. This causes problem because we can have READ_ONLY nodes that are not part of the rebuilding set (e.g. with MAY_DISAPPEAR maintenance). This diff makes NCM the source of truth for whether the shard is writable or not.

Note: Along the way, this diff removes one more callsite to `Worker::onThisThread` by keeping an updatable NodesConfiguration instead of asking the worker about it.

Reviewed By: tau0

Differential Revision: D22645904

fbshipit-source-id: ea76f6437823101283c756c8bdbc5a9c0e416b39