EpochStoreV2: Unify the logic between ZookeeperEpochStore and FileEpochStore
Summary:
For some reason, the epoch store request logic was duplicated between ZookeeperEpochStore and FileEpochStore. This diff changes that by modifying FileEpochStore to utilize the same zookeeper epoch store requests.

Note: It's not optimal that the requests are called `ZookeepeerEpochStoreRequests` while being used in FileEpochStore but I'll change that later.

Note (2): FileEpochStore is going to completely die when we migrate to the `VersionedConfigStore` interface, but this requires the new thrift-based serialization format. So it lives till then.

Note (3): There's still some duplication in the way we handle the result of the request, will leave it like that for now as I'll be touching it again when I'm doing the double writes.

Reviewed By: 747mmHg

Differential Revision: D23218240

fbshipit-source-id: 790a85f84e5c862ee775b55d7d71c4b9c4c451c4