EpochStoreV2: The new serialization schema for EpochMetadata
Summary:
The new thrift definitions of epoch store. It defines a new `LogMetaData` structure that's used to store what currently is stored in `EpochMetaData`, LastCleanEpoch information and `EpochStoreMetaProperties`. This is going to be the data that gets stored in epoch store for each log.

Note (1): I decided not to create a structure for LCE, because the metadata log part is going away at the end of this project.

Reviewed By: rashedrahman, 747mmHg

Differential Revision: D23393943

fbshipit-source-id: 3fc14b795af64d0d5489c11c85edeb511572e1cf