Merge connection iteration methods
Summary:
We have 2 very similar methods: `forEachConnection` (1) and `forAllClientConnection` (2). The only difference is the second one skips server connections on iteration.
Given that `Connection` object has enough information for checking whether it is client or server connection and that (2) is not used on critical path (only in single test and one admin command) lets merge these functions and simplify interface.

Reviewed By: MohamedBassem

Differential Revision: D23344686

fbshipit-source-id: ec0754e7ed88fdd6f8411a120b0f44e2228d1679