Move peer_node_id and peer_type into ConnectionInfo
Summary:
They now will be set and read through `ConnectionInfo`. It may look wasteful to set the whole object in handshake phase to update single field but later I will add more fields and most of the structure will be updated at the place.
I also made peer type to be calculated on-the-fly as it fully defined by other fields, otherwise we need to apply extra efforts to ensure these fields are in sync.

Reviewed By: MohamedBassem

Differential Revision: D23448570

fbshipit-source-id: 91fca540c50ea727ee28a2b43b6e7d21e1880ce4