EpochStoreV2: Change interface of ZookeeperEpochStoreRequest to use LogMetaData
Summary:
In preparation for the double writes in the next diff, we need to change the interface of ZookeeperEpochStoreRequest. The current way it works is that each request deserializes and stores the important information into member variables. This needs to change to explicitly provide the structure that the request will be working on, so those changes were done:

1. The `onGetZnode` function is now split into two functions. One that deserializes the legacy znode into a partially filled `LogMetaData` and the other one is the one that contains the actual logic.
2. Requests are no longer allowed to persist any of the deserialized data that's why the deserialize function is const qualified.
3. Given that we no longer persist information about the data the request is working on, `LogMetaData` needs to be passed to both `composeZnodeValue` and `postCompletion`.

The diff modifies the 3 requests to implement the new changes to the interface. The diff also modifies both ZookeeperEpochStore and FileEpochStore to implement the new interactions.

The way I suggest reviewing this diff:
1. Understand the changes done in the interface in ZookeeperEpochStoreRequest.
2. Review the changes in the 3 requests that implement that interface.
3. Review the changes done in ZookeeperEpochStore and FileEpochStore.

Reviewed By: 747mmHg

Differential Revision: D23576084

fbshipit-source-id: a6f0afbac844c48933c26f04ca564c3eac25f56a