ClusterState to set nodes to UNKNOWN by default on insertion
Summary:
ClusterState will initialise newly added nodes as either `FULLY_STARTED` or `DEAD` depending on whether failure detector is running or not (are we a server or not).

This diff introduces a new state `UNKNOWN` that will be set only if failure detector is (not) running (client-side) to avoid making this a breaking change. The UNKNOWN state can be masked to any other default state as requested by the user in most of the calls to get the state. The default in this case is to assume that nodes are FULLY_STARTED. This means that this is 100% compatible with any old code but in the case of Maintenance Manager, we care about this and the default will be set to DEAD.

Reviewed By: mcrnic

Differential Revision: D23471431

fbshipit-source-id: a59072a3cc54ae9624173d51eff3fdfa7630bd26