Print task and container handle of node with higher version
Summary:
This should make it easier to debug issue.

Node with higher version could be from other task (or same task).   For other task, it is likely mean partitioning. For same task, it means that version goes backward.

Also print container handle  since DPDaemon asssignment key is based on it

Reviewed By: mcrnic

Differential Revision: D23386552

fbshipit-source-id: 9daf78b07cecc382f4b27cc89b614b44c24d509a