Get TW task handle for DP nodes
Summary: When getting a list of nodes the tests assume that the given `NodeView::node_name` is a TW task handle. On DP clusters the DP id is returned as node name (e.g. `DPID_1:39`). However, DP also sets the additional `NodeConfig::tags["handle"]` with the correct TW task handle.

Differential Revision: D23846035

fbshipit-source-id: d93b8b3cea14918b340206958bda0d2e7d1fc4d1