EpochStoreV2: Deserialize the migration znode LogMetaData
Summary: Now that we read the value of the migration znode, let's start deserializing it into the `LogMetaData` structure. Given that we still want to legacy znode to be the source of truth, the deserialization of the legacy znode happens after deserializing the migration znode and overrides the values the requests care about.

Reviewed By: 747mmHg

Differential Revision: D23690297

fbshipit-source-id: 7b802ec9357e0ce5ba4a4a034df2ef8274f0caee