Removing extras #1: Change unit tests for Appender
Summary:
== Removing extras ==

This stack of diff is about removing extras in the code base.

=== What is extras? ===
"extras" is a feature that consists in speculatively selecting more than R
(replication factor) recipients in the copyset for an APPEND. This allows more
predictable latencies of appends as Appender just needs to wait for the first R
copies to acknowledge the write before retiring. As an optimization to save
storage resources, Appender would send DELETE messages to unused recipients once
it's successfully replicated the record, without waiting for these DELETEs to be
acknowledged.

This feature is configured by the "extraCopies" attribute in logs config. This
is not to be confused with another piece of metadata called "extras" which
allows passing arbitrary KV metadata to a log group (such as oncall for
example).

For example, if replication=3 and extras=2, Appender sends 5 stores and will
acknowledge the APPEND to the client once any 3 recipient has acknowledged. If
there is notification of a failure to store a copy on any recipient, Appender
checks whether there are enough remaining outstanding stores in flight such that
we can hope to complete the write. If that is not the case, Appender starts
another wave. Without extras, Appender basically retries a wave as soon as any
recipient fails the store.

=== Is this used in production? ===

No. It cannot be used in production for several reasons.

1/ Extras make rebuilding and Single Copy Delivery unsafe. Indeed, if
replication=3 and extras=1, effectively the serialized copyset in local log
store will have size 4. When a storage node reads this copyset, it's unclear
which recipients in the copyset actually have a copy or not. Both Rebuilding and
Single Copy Delivery rely on the copyset to make decisions about which recipient
should do work. Extras make this logic more complex and error prone. One could
modify Appender to have a final stage that sends a final AMEND message to all 3
copies to remove the "extras" from their copyset, but again this is additional
complexity and would not be 100% safe if the Appender crashes before this stage.

2/ It's harder to implement copyset selection. Since we stopped using extras in
production because of 1/, we also implemented new copyset selectors that have to
navigate many constraints such as replicating over failure domains.
These selectors do not work with extras and actually silently discard your
intent to use them if you were to configure a log with that feature.

=== How is extras adding complexity in the code base? ===

This is a setting that adds a fair amount of boilerplate code in copyset
selectors, logs configuration and tests. It is part of our public documentation
although unusable. It is one less thing to wonder/worry about.

By the way, I saw we also have a "syncedCopies" option that I'm pretty sure we
also don't use in production.

=== Is there a future where we do have something similar? ===

In a world where we build a digest-based read path (Read Path 2.0) and
centralized rebuilding also based on digest, copysets will not be useful
anymore. This would remove problem 1/.

Problem 2/ would still exist. Some systems speculatively attempt to select an
additional node if the current time that has passed is greater than say the 99th
percentile of observed latencies in past requests. In a world where we don't
have copysets (and therefore don't need to AMEND and resend 3 STORE messages in
each wave), this process would be there for free thanks to the "wave mechanism"
that we already have in Appender.

== Unit tests for Appender ==

More specifically, this diffs removes any logic in the unit tests for Appender
related to verifying the behavior when extras are enabled.

Reviewed By: MohamedBassem

Differential Revision: D23867744

fbshipit-source-id: ce188095cf6ac30e454649455675344471c4db5c