Removing extras #3: Modify CopysetSelectors
Summary:
Read the summary of D23867744 for a full explaination of what we are
doing here.

In this diff I am modifying the APIs of CopysetSelector and its derived class as
well as CopySetManager and it's derived classes to remove the notion of extras.

This is diff is pretty much a no-op, the most recent copyset selectors did not
implement support for extras anyway (were returning E::PARTIAL if extras are
configured).

Depends on D23867746

Reviewed By: MohamedBassem

Differential Revision: D23867743

fbshipit-source-id: d57d3a62147616bf3279cf536013b768d76bdad8