Removing extras #2: Modify Appender and RecipientSet
Summary:
Read the summary of D23867744 for a full explaination of what we are doing here.
In this diff, I am modifying Appender and RecipientSet to remove the extras
feature.

This is mostly boilerplate code. Appender uses a CopySetSelector that may return
E::PARTIAL when extras are used, but since extras is overidden to be zero when
using the selector, we are asserting that we never get this result.

In a later diff, I remove the "extras" parameter from the API of CopysetSelector
anyway.

This removes the phase in Appender after replication that sends a DELETE message
to all outstanding recipients, since we know this code will never be useful
without extras.

Depends on D23867744

Reviewed By: MohamedBassem

Differential Revision: D23867746

fbshipit-source-id: 8599c008047ea0f91a9d5938f4abfa316f2a3d23