Migrate to new AsyncSSLSocket session API
Summary:
Migrate logdevice to the new AsyncSSLSocket session API, in preparation for enabling TLS 1.3. The new session given by AsyncSSLSocket will be opaque and will not allow access to the internals of OpenSSL's SSL_SESSION. SSL_SESSION was being used for a few things:
1. Using session id compare sessions, which can now be done by comparing `shared_ptr<SSLSession>`
2. Checking if the session has a ticket, which should not be necessary in TLS 1.3
3. Checking the lifetime of the session. This should not be necessary, since it should fallback to a full handshake if the session has expired. This will affect the stats, since session resumption is considered to have been attempted when a session is retrieved from the cache, but the difference should be slight.

Additionally, session caching on SSL_CTX has been un-disabled, as it is required for obtaining sessions in TLS 1.3.

Reviewed By: MohamedBassem

Differential Revision: D22023139

fbshipit-source-id: d663e7c5c81a413e25a3d8d77e4b609f23c4a9a0