Updates NodeSetTest and Fixing a bug in selectors donors for rebuilding
Summary:
This diffs migrates the NodeSetTest integration tests to use the standalone admin server and trigger maintenances instead of manipulating eventlog directory to trigger rebuilding.

The change uncovers a bug that we have seen in production where rebuilding can get stuck due to one node being expected to be a donor but it doesn't know that and it never rebuilds the data.

The correctness issue is related to how we have a dependency between the RSM state (EventLogRebuildingSet) and the state of NCM (NodesConfiguration) at the time of replay. Nodes that are in state NONE are discounted from being a potential donor (they will be excluded already if they are AUTHORITATIVE_EMPTY) and that node may get out of this state later causing it to be considered donor as they process the `SHARD_UNRECOVERABLE` message. The node itself will not re-evaluate the authoritative status at this point and will not trigger rebuilding.

Proposed solution:
- Make the RSM replay more independent from NCM's current state except if the node doesn't exist in nodes configuration altogether.
- Tests now pass reliably.
- Add more tests to cover more edge and complex cases (next diffs)

Reviewed By: MohamedBassem

Differential Revision: D24044152

fbshipit-source-id: 11d4ab966115f77fd3e79b017c0e3a16abfd114e