fb/py/ldops/hardware.py: change MEDIUM port
Summary:
Expand handlers, cluster turnups, etc will now use a different MEDIUM port as
in a previous diff we started creating listeners for the MEDIUM port (which
means that if MEDIUM port is the same as default_data_port, there will be a
clash).

To be rolled out some time after the previous diff is committed, when
tooling/daemon/etc catch up.

Reviewed By: MohamedBassem

Differential Revision: D23813406

fbshipit-source-id: 51285ecf316cbb6478fb03fa635ad799c93f7eee